[INTERNAL][FIX] sap.ui.fl Fix issue of uninitialized FlexState
ControlPersonalizationWriteAPI functions can be called for a component
that has no changes yet. So it is possible that the FlexState of the
component is not yet initialized before changes got
add/save/restore/reset. That could lead to exception. Therefore,
FlexState need to be initialized before any update action could be done.

Change-Id: I4139638c808889dbec245f4de8164c485cfb41c8
BCP: 2280091138