[INTERNAL][FEATURE] TreeBindingProxy: Provide proxy for TreeBindings
To enable the usage of OData V4 in the TreeTable and other tree controls, a proxy was implemented providing a central place for handling the different behaviours of OData V4 and older bindings.
The TreeBindingProxy provides an API, which allows to call "binding" methods directly without worrying about the binding type.

JIRA: CPOUIFTEAMB-2049
Change-Id: If282451487eaf5ff97b990bd65aaaae9ba705394