[INTERNAL] sap.ui.core.mvc: Fix expectations of JSONView unit tests
These tests would fail, if the sap.m.DatePicker introduces a new
property. This would lead to an additional extractBindingInfo call,
which is the correct behaviour. Hence, removing
the 'callCount' assertions.
The intention of these test is to test the returning value of
the function for aggregation objects with value property.
This part has been covered.

Change-Id: I2c30ad32aef51adfbe6568180f1a8f3ed26a2809