[INTERNAL] sap.uxap.ObjectPageSectionBase: Improve accessibility for hidden title
Problem: The title of the ObjectPage(Sub)Section was still mapped to the accessibility tree even if it was hidden.
With promoted sections with hidden titles, this was causing the screen reader to read the title of the promoted section twice.

Solution: The title is now only mapped to the accessibility tree if it is visible. This is done by excluding its value from the the aria-labelledby attribute.
The getShowTitle method added to the base class is to safeguard the implementation as the property is available only on SubSection and Section level.

Change-Id: Ifbbf8f9d21d2fc2fb128a35b45c8a820c84640e6
BCP: 2280163428