[FIX] sap.m.IllustratedMessage: Fixed vertical responsiveness
Problem: When 'enableVerticalResponsiveness' is set to 'true', sometimes
the content of the IllustratedMessage is going out of the bounds of its
parent container.

Root Cause: Recent regression where the DOM structure of the
IllustratedMessage was altered.

Solution: We make sure to apply the correct CSS styles to the newly
added DOM element in order to achieve the target design.

BCP: 2270144551
Change-Id: I7ed481b548989fdf2f43219e55bbb9657dd63d25