[INTERNAL] OPA: Enable check for Component parents
Problem: When Control is placed in a Component, which is placed in
Control (child of another Component), OPA matchers does not check the
whole parent-tree, but stops at the first met Component (from
inside-out). This way, for example, if there is a Control, placed in a
Component, placed in a VBox, which is currently busy, this Control will
be treated as interactable, while it should not be.

Solution: Check the whole parent-tree by not stopping when meeting a
Component, but instead - go up the parent-tree.

JIRA: BGSOFUIPIRIN-6001

Change-Id: Ie7c7e7766740b1d864b545376bf2d486eac5111a