[FIX] sap.m.Avatar: Avatar initials fixed
Problem: Initials of the of sap.m.Avatar are not displayed in some cases.
This happens, because we check the scrollWidth of the initials,
to see whether they are overflowing the avatar (this check is needed
only when there are 3 initilas inside the avatar). By the time, we
check the scrollWidth, the css styles are still not loaded in the
testsuite and this leads to the case where the scrollWidth is 0.

Solution: We could isolate and check for the scrollWidth only when
the sap.m.Avatar has 3 initilas inside - anyway, this is needed
and will fix the failing visual tests.

Change-Id: I237410a10baaf0d869418377d44169697fae51c6