[INTERNAL] sap.uxap.ObjectPageLayout: Do not hide focused elements
Problem: When an element receives focus and it is under the floating
footer, the page is not scrolled and the element stays hidden.

Solution: Do not allow the ObjectPage to scroll to the bottom of the
visible content, but instead - allow it to scroll only to the top of the
floating footer, using scroll-padding-bottom CSS property. This way, if
an element is currently under the floating footer and it receives focus,
the Browser tries to get it into view and scrolls up the floating footer
(due to the scroll-padding-bottom), which limits the scrollable view
port height.

JIRA: BGSOFUIPIRIN-5865
Change-Id: I2436e0b95b4de986f3a8909eec014f2ff47ea288