[FIX] sap.m.Avatar: Missing initials fixed
Problem: Initials of the of sap.m.Avatar are not displayed in
some cases.
This happens, because we check the scrollWidth of the initials,
to see whether they are overflowing the avatar. By the time, we
do it, the css styles are still not loaded in some environments
and this leads to the case where we do not get the correct width.

Solution: Instead of onAfterRendering, we could check the width on
onThemeChanged, in order to be sure that we have all styles applied.

Change-Id: I03e22ab6f5c04d565536f7047fbf2c4421da725f