[INTERNAL] sap.ui.dom: introduce new utility isBehindOtherElement
Problem: There are some cases, when the focus may be moved on an
element, which is behind another element. For example, when a
MessagePopover is opened and we click on an item with an error and move
the focus on the corresponding Input.

Solution: Introduce new utility - isBehindOtherElement method, which
checks if the element of an interest is currently behind another
element. If so, the application could either scroll the page or close
the MessagePopover.

BCP: BGSOFUIPIRIN-5866
Change-Id: I1392ddb66a74bcec767a82afc6e7177c57a5c6b8