[INTERNAL][FIX]sap.ui.rta: Improve documentation links
The "Define Embedded Content" (add iFrame) dialog and the warning
message that is shown when there are changes from different systems in
the same app have direct links to a particular product and version.

To ensure that the key user gets to the proper documentation (both the
right product and up-to-date information), we replace the direct links
with the help portal search results for the corresponding topics. From
there they can choose their product.

Long-term we should detect the proper product/version and have a direct
link - but currently there is no utility available to retrieve that.

Change-Id: Iaa760c568099ccdf37bca1c6aec59b160f22b0b3
BCP: 2270144233