[INTERNAL][FIX]sap.ui.rta: AddLibrary for lazy load
When an app descriptor change "addLibraries" is created, we only check
if the library is already part of the manifest dependencies or not. In
cases like Add Field, if these libraries are loaded with "lazy:true", it
causes issues due to the delegator not being loaded before the change is
applied.

This correction ensures that "addLibraries" will be called also for
libraries which are part of the app dependencies with "lazy:true", which
will change the entry in the manifest to "lazy:false" and prevent the
timing issues.

Change-Id: Iead3f05905233375ef76fdfd1c936a96faf1022c
BCP: 002075129400007205882022
CR-Id: 002075125800004637212022
(cherry picked from commit b776264776b1a9d3aa780ce1cc7b33a650ab49c6)