[INTERNAL] Field: with unit or currency wrong error message
If a wrong number or unit/currency is entered the shown error message might be misleading. Here only the error of the currently used part (number or unit) should be shown.

It is possible to enter a number in the unit/currency part. There only units or currencies should be allowed.
Move creation logic for unit/currency type to TypeUtil as it might be type dependent.

Change-Id: I95ac39afbdb0200a320201d9bc164b16cc2de81a
BCP: 2270153230