[INTERNAL][FIX]sap.ui.rta: AddLibrary for lazy load
When an app descriptor change "addLibraries" is created, we only check
if the library is already part of the manifest dependencies or not. In
cases like Add Field, if these libraries are loaded with "lazy:true", it
causes issues due to the delegator not being loaded before the change is
applied.
This correction ensures that "addLibraries" will be called also for
libraries which are part of the app dependencies with "lazy:true", which
will change the entry in the manifest to "lazy:false" and prevent the
timing issues.

Change-Id: Iff39d6e7e5fedbad13d342737bac1dd203540044
BCP: 002075129400007205882022
CR-Id: 002075125800004636652022