[INTERNAL] sap.ui.core.ScrollEnablement: Skip option provided
New option is added to the scrollToElement API method of the
sap.ui.core.ScrollEnablement. If the new bSkipElementsInScrollport parameter
is set to true, that would cause scrolling to happen only if necessary,
i.e only if the element is not already visible in the scroll container
view port.

JIRA: BGSOFUIPIRIN-5901
Change-Id: I96b29c54c29ac78aeaefaf6a91167605d883530d