[INTERNAL] MDCTable: Add propertyInfo property
The purpose of this property is to get faster to the initialization and the initial data request. The property has to contain all information that is needed for this. If additional information are required during runtime, the property infos are fetched (finalized) from the delegate.

Properties are finalized when:
 - P13n dialog is opened in runtime
 - P13n dialog is opened in RTA
 - Column menu is opened
 - Changes are applied (Variant switch, StateUtil)
 - Export is triggered
 - Calculating the width of an added column whose dataProperty is not available

If a column is added, the properties are not loaded. It is the responsibility of the user to load properties in TableDelegate.addItem if required.

Updates when the properties are finalized:
 - PropertyInfos of the GridTable ODataV4 plugin inside odata.v4.TableDelegate
 - If property finalization was triggered by a flex change, rebind is deferred until finalization is completed

Change-Id: I96d2b48179f5cf613b19b7c2d232a90dc3ac0709
Jira: CPOUIFTEAMB-1932