[INTERNAL] FilterField: use binding to parse user input
use the same concept to parse user input on multi-value FilterFields like for single-value.
Bind the "value" property to "conditions" using ConditionsType. But prevent the formatting of the single conditions into the "value" property of the MultiInput.

For SearchField use the binding to update the value if the change-event is fired.

So all special logic to parse values outside binding can be removed

Change-Id: I0314e37c17065a4577d6798d2a87b9b804024b3a
JIRA: CPOUIFDENVER-1842