[INTERNAL] MTable: do not focus table in single-value navigation
If a ValueHelp using a MTable is used as typeahead for a multi-value field the table could be configured as singleselect or multi-select. If multi-select is used, the focus needs to move to the table on arrow-key navigation to choose multiple values. If single-select is used the focus should stay in the field using arrow keys and the navigated text needs to be taken into the field.

Change-Id: I8a28a01fb177a5a439b50f86ad7c5dbe9d867db8