[INTERNAL]sap.ui.rta: Switch on modified Smart Variants
Modified (dirty) variants can be switched to another variant without
the user knowing if the changes were saved or not. This change improves
the behavior so that, if there are changes, a dialog is displayed to
the user asking if they want to save or discard the changes; or cancel
the variant switch operation.

Change-Id: If2387fb63a37ee1bd1a579243f7a756bd436aa2c
JIRA: TEAM42-229