[FIX] sap.ui.core.date.Japanese: calendar week
Fix the calendar week calculation, such that the Japanese calendar
uses the Gregorian year. This is done because the Japanese emperor
year depends on the era and is not unique.

Before:
Without era, the current era was used (Japanese constructor), so every
week calculation was done using the current era.

Now:
The Gregorian year is used and the week calculation is correct again.

Note: The constructor of Japanese is able to handle the gregorian year (>100)
Partial revert of: 5643018

PS1: cherry-pick of 5750280
PS2: review comments (formatting)

BCP: 2270172759
CR-Id: 002075125900003474622022
Change-Id: Ib0019ab167fae09d4d6f3ef68c74a81710d99d0f
(cherry picked from commit ea3e28b57118b2609cd9dce38cb1681484eb6893)