[FIX] sap.m.Input: Typе-ahead now selects the first matching item from the suggestion list or table
When the suggestion items/rows were not initially available
but were presented on a later stage (e.g. suggest event),
the type-ahead functionality was only extracting the text value
from the item, but the item itself was left unselected.
This is no longer the case and if an item was chosen and the
user's input was auto-completed with its text, this same item
will become selected in the list.
This hoewever will not be reflected in the selectedRow/Item
associations in the Input, untill the interaction is finished.
A sample containing two examples - one with list suggestions
and one with tabular was added as well (InputDynamicSuggestions).

Change-Id: I685a1c748c23cf830514189f529f53f93436f1f9
BCP: 2270141561
BCP: 002075129400001470182023
BCP: 002075125900001634732023
CR-Id: 002075125900001634732023
(cherry picked from commit af2d351f7ef276d4d3cdab0ae679abc1ea890dbf)