[FIX] sap.m.ProgressIndicator: Visual misalignment of percent value fixed
Problem: When the table is scrolled down, its rows are reused and all
controls, placed in them, are also rebinded with new data. When
ProgressIndicator has a percentValue greater than 0 at first, then it is
set to 0 and the ProgressIndicator is removed from its parent (not
rendered), and after that the percentValue is set to something greater
than 0 and it is added to its parent's aggregation, the "sapMPIValueMin"
class is not removed. As a result, the percent value part of the
ProgressIndicator looks broken.

Solution: Update CSS classes of the ProgressIndicator instance, no
matter whether it is rendered at the moment or not.

BCP: 2270167989
Change-Id: If4a968afa545afce70affbd0adfc9cfeb18c5c67