[INTERNAL][FIX]sap.ui.fl: Missing author after navigation
In FL Variants, if you create a new variant ("save as") and navigate
away from the app (e.g. back to FLP Home), then go back to the app, the
"author" field for that variant will be empty in the "manage" dialog of
the Variant Management control.

This happens because the author is first added to the variant after it
gets persisted (field "user" from SupportInformation), and on creation
we directly set the author in the variant so it is immediately visible.
However, when we navigate away and back, the model is not updated with
the user information retrieved from the backend - and the author field
is gone.

This fix checks if there are variants without user when building the
variant map and adds the current user (from Settings). Since this
intermediary state is only possible within the same session, we can be
sure that the current user is also the creator of the variant.

BCP: 2270154387
Change-Id: I5bf95f0fbfc421435c6f60425089c2f81637f781