[INTERNAL] MDCTable: Refactor table types
Changes all static methods on TableTypeBase and subclasses to instance methods (except createColumn for now). This should allow to easier create further types, for example for TreeTable support.
Table#_getStringType and Table._bMobileTable are replaced with Table#_isOfType. This method allows checking for a concrete type and include subclasses or not.

Change-Id: I2d0c0c8af3b6c273f63f0297a3b8c8a9c9be9b5f