[FEATURE] sap.m.OverflowMenuButton: New control for OverflowToolbar
Issue:
MenuButton is not consistent with other controls in the toolbar overflow.

Solution:
Introduce new control OverflowToolbarMenuButton that should be used in the context of anOverflowToolbar. Similarly to OverflowToolbarButton and OverflowToolbarToggleButton, the control displays only an icon in the toolbar and an icon with text in the overflow area.

Included an sample of the component as a part of the "OverflowToolbar - in a table and in the footer of a page" sample.

JIRA: BGSOFUIBALKAN-6403
Change-Id: I81c2dbed26bf73115c7e426d293bcea57d056e13