[INTERNAL]sap.ui.rta: Switch on modified Smart Variants (V4)
Modified (dirty) variants can be switched to another variant without
the user knowing if the changes were saved or not. This change improves
the behavior so that, if there are changes, a dialog is displayed to
the user asking if they want to save or discard the changes; or cancel
the variant switch operation.

JIRA: TEAM42-230
Change-Id: I8bbcf74b9b4d8b03bd993250f151eed7d470cfca