[INTERNAL] sap.ui.rta: fix modified flag on CompVariantContent undo
To set the correct modified flag after the undo of an CompVariantContent command
the flag has to be fetched before and saved in the command.
But at the point in time where it was fetched a control already changed it,
so it is now retrieved before the dialog gets opened.

BCP: 2270169098
Change-Id: Id3e1fa4c39ccbf99508bcd313e03413d087181b0