[INTERNAL][FIX] Table: Fix Column Width Calculation when column contains mandatory content
In a previous change, the column width calculation was enhanced to take the "required" state of a column header into account.
The proposed change has only effects on the MDCTable and does not apply to the SmartTable.
Therefore this change introduces the calculation utility in the SmartTable as well, by implementing the functionality in Util#calcHeaderWidth.

Change-Id: Ie0af9f144585046d7e4b73da4ee63c4e48e14f3b