[INTERNAL] Filter settings enhancement
The ExportHandler provides enhanced filter
settings on its beforeExport event. Those
settings will be used to update the labels
and types of the filter settings so that the
generated information in the document
matches the user expectation.

The new event parameter on the beforeExport
event of the ExportHandler, will also be
available in the beforeExport event of the
mdc.Table. This allows subsequent layers to
further enhance the filter settings.

JIRA: CPOUIFTEAMB-2054
Change-Id: I3eb1f5e2af312cf5dd7ca56a27c0d9aa95039e7d