[FIX] sap.uxap.ObjectPageSubSection: fix RTA Remove action
Problem: When there is only one SubSection in a given Section, the RTA
Remove action should not be enabled and should not be able to be
executed on the SubSection. Now this functionality is not working, as we
have an aggregation forwarding for the "subSections" aggregation of the
Section, as the .getParent method is overriden (for backward
compatibility), but the sParentAggregationName property cannot be
changed, because it will break our new functionality.

Solution: Using the isEnabled function in the designtime metadata of the
SubSection, to check if the SubSection is the only one (visible) in its
parent Section.

BCP: 2270127057
Change-Id: I595a16a41d4122cfdce4469bafff9a180132b553