[INTERNAL] Demo Kit: Removed button for theme changing
Problem: The theme for the samples can be changed in two ways: via the
left-hand side toolbar menu button and from the right-hand side toolbar
of the actual control sample. The latter is the preferred way, as it
allows changing the theme only for the sample and not for the whole
page plus the sample, which is confusing.

Solution: Removed the theme-changing button from the left-hand side
and all its related functionality

JIRA: BGSOFUIPIRIN-5907
Change-Id: Idabe4d8d42b4f5a6e0ca919f914c10b05ff89b6a