[FIX] ManagedObjectObserver: performance optimization for destroy
Introduce new shortcut (WeakMap) from a Listener to its Targets.
This improves the destruction logic as we do not need to iterate
all Targets to find a Listener, but only iterate the necessary
targets.
In Applications heavily use nested FlexBoxes (HBox, vBox) the amount
of targets is quite huge. This leads to severe performance degradations.

Change-Id: I3f139acbfda8899b766c6663416a4e00cff11f31
BCP: 002075129400008610112022
CR-Id: 002075125800004880312022