[INTERNAL] sap.m.Popover: Visual tests enhancements
- Previously, we were changing the themes between SAP Quartz Light
and Belize but that does not make sense anymore, as it was done
in order to show the responsive paddings in Fiori when the tests were
executed on Belize by default.
Now, the tests are executed in Horizon by default and the responsive paddings
are presented there, so these use cases are not valid anymore.

Change-Id: I0479696c3fdec8051fe578582f3e14d2d2a33a26