[FIX] sap.ui.test: Artificial focus events complemented
Problem:
Missing "focusout"/"blur" events when shifting the focus
to a new element.
This happened on Firefox where OPA simulates the focus
events and fires only "focusin" and "focus" upon focusing a
new element and skips "focusout" and "blur" on the previously
focused element.
In some scenarios the artificial "focusout"/"blur" events
are fired by default in a previous step
(from sap.ui.test.actions.EnterText when its "keepFocus" property
is disabled). However, in those cases the
"focusout"/"blur" events are called *too early* which gives unexpected
results (compared to the scenario with native events) in tests,
as in the OPA test from the linked BCP ticket.

Solution:
Ensure the "focusout" and "blur" [on the previously focused
element] are also always automatically fired when simulating
focus shift to a new element.

Change-Id: Id3672c08318796c748ec1be0968960d4d6477a28
BCP: 2270156201