[INTERNAL] sap.uxap.ObjectPageLayout: optimize lazy loading
Problem: When there are relatively small (in height) SubSections,
especially the first two, the LazyLoader unstashes unnecessary huge
number of SubSections. This happens because ot two reasons:
1. The first two SubSections (unstashed before the first rendering of
   the OPL) take small part of the screen, and as all of the other
   SubSections are stashed, they take 16 or 64 px height. In this case,
   (almost) all SubSections are getting unstashed.
2. The LazyLoader has functionality to unstash an extra one
   SubSection (preloading), if it's positioned less than a screen and
   a half away from the current scroll top position. As FE call a few
   times sync _triggerVisibleSubSectionsEvents - an extra SubSection
   is getting unstashed for each call.

Solution: Set dummy height (20% ot the view port) to the stashed
(empty) SubSections, before being unstashed. This way we prevent the
unstashing of too many SubSections at a time. 20% height is considered
better than, for exmaple, 50%, as with 50% we risk to unstash
SubSections on two rounds (async), which may lead to worse performance
(especially if there are SmartTables inside -> there is possibility of
additional batch requests).
Also removing the preloading functionality of an extra SubSection, as it
adds time to the loading of the OPL (which always takes more time),
saving just a small amount of time, when user scrolls down.

BCP: 2380123675
Change-Id: I95b6327e33b26d5bce73d8a5cb7c48521f9f3f09
CR-Id: 002075125900003140402023