[INTERNAL] sap.ui.fl: add custom handling for FlexCustomData
During XML preprocessing the flex custom data are written as namespaced attributes.
Retrieving this custom data can be done synchronously, but instead the generic
'getAggregation' of the modifiers was used. This had some special handling
where it creates custom data objects for all the namespaced attributes.
This is an unnecessary overhead in the case of FlexCustomData.
Performance will be drastically improved for XML preprocessing with high amounts of changes on the same control.

Some synchronous extra functions removed since the formerly async function is now also sync
	- FlexCustomData.sync.getAppliedCustomDataValue
	- FlexCustomData.sync.getParsedRevertDataFromCustomData
	- FlexCustomData.sync.hasChangeApplyFinishedCustomData
	- FlexCustomData.sync.destroyAppliedCustomData
	- apply/_internal/changes/Utils.checkIfDependencyIsStillValidSync
	- ChangePersistence.copyDependenciesFromInitialChangesMapSync

JIRA: TEAM42-235

Change-Id: Ifd03a3436b54186038366b29618539d063d9067a