[INTERNAL] sap.ui.rta: add new RuntimeAuthoring API that includes condensing
Currently the Visual Editor uses a private API to save.
This is also used to condense, controlled via a parameter.
In the future the relevant changes need to be passed to the API, so a new API is now introduced.
For now the function simply calls the save API, so no change yet.

Change-Id: I32db9c22f13823fba770d6899d28d8b2ae42602a