[INTERNAL][FIX]sap.ui.rta: Avoid scrolling when CViz starts
Change indicators get focus when Change Visualization starts. If the
application is scrolled down and the first indicator is on the hidden
area, this causes an unexpected scrolling to focus the first indicator.

This fix gives focus to the first indicator that is visible with the
current scroll situation, preventing unwanted scrolling when switching
to Visualization mode.

Change-Id: I4fa11b0468b3291c0bbe3e2e2bbbf145af6498b4
BCP: 2270178026