[FIX] sap.uxap.AnchorBar: Selecting Section with keyboard
Problem: When users select Section/SubSection from Hierarchical Select
with the keyboard, the selected Section/SubSection is not focused.

Solution: Previous fix, which focuses the selected Section/SubSection
from the AnchorBar, was implemented, when the user makes a mouse click.
However, it seems that when he navigates with the keyboard, the problem
still exists. It comes from the onkeyup event of the Select, which closes
the Popover, and upon this closing, the Popover applies focus on its
opener (the arrow down). This happens after the AnchorBar moves the
focus on the selected Section/SubSection. That's why now the AnchorBar
moves the focus with a delay, to make sure it happens after the focus
restore of the Popover. The QUnit is also adjusted, in order to
demonstrate the real situation with a keyboard navigation.

BCP: 2270113573
BCP: 2370011663
Change-Id: I223750ecd3f91616f30b4f86529ce004d465c54a
(cherry picked from commit 318a762c85839cdc97c4db158f994b3590ba90db)
CR-Id: 002075125900000532112023