[INTERNAL] sap.f,sap.uxap: Remove obsolete scrollToElement callbacks
The recently added ScrollEnablement.prototype.setScrollPaddingTop
now makes unnecessary the use of callbacks for end of scrollToElement
(that we earlier used to attach via
ScrollEnablement.prototype.setOnAfterScrollToElement
with the purpose of offsetting the scrolled element in some cases).

This change deletes the old listeners and instead makes
ScrollEnablement.prototype.scrollToElement take the
required offset *in advance* (where the offset value is the value of
scrollPaddingTop) for computing the final position to scroll to.

JIRA: BGSOFUIPIRIN-5901
Change-Id: I446e9870775b2c65f652f549b49afcbfc2b0ad10