[INTERNAL] p13n: improve 'labelFor' handling
This change fixes the 'labelFor' reference handling for the controls:
- GroupView
- SelectionPanel
- FilterPanel

The issue is related to the additional wrapping FlexBox in some implementation
scenarios. With this change, the #getAccessbilityInfo method is being
used, which may return the acc relevant children, which is the according
input control for instance in the mdc library usage.

With this change, the Table & Chart 'View Settings' Dialog items in the
filter tab can be clicked on the text to jump into the filterable control.
The same is now also working for the 'Adapt Filters' Dialog - this now
also works across V2 & V4.

BCP: 2270188480
BCP: 2180253944
BCP: 2270149967
Change-Id: Ie3e96c347a16aa089afc2346dcb1871b5490326e