[INTERNAL] TreeTable/MDCTable: Disable column reordering for Tree Column via Drag&Drop
The first column of the TreeTable should not be movable by any means. This includes drag and drop and keyboard shortcuts.
This change makes the first column non-draggable and therefore disables the possibility to reorder the column via drag& drop.

Feature enhancement is hidden behind a URL parameter (sap-ui-xx-v4tree-p13n) for testing purposes, as hiding the first column is not possible yet (CPOUIFDENVER-1861).

JIRA: CPOUIFTEAMB-2082
Change-Id: Ib897b289fb7b642a1a9d9c5b10ac735f4dbcabe9