[FIX] sap.m.MultiComboBox: Selection in tokenizer is preserved after opening the picker
Previously, after the picker content was synced, the tokens were
created in order corresponding to their matching list items from
the MultiComboBox's items.
Now the tokens are created in the same way the user has selected
them or they were provided via the selectedKeys property or the
selectedItems association.
BCP: 2170315279

Change-Id: Ifc22a50adfb147b730d07c2e1499b400a987538e