[FIX] sap.m.OverflowToolbar: fix error on opening Popover
Problem: ToggleButtons are invalidated onBeforeRendering, which
causes an error, when trying to open the OverflowToolbar's Popover menu.
This leads to resetting of the OverflowToolbar and rerender. As a
result, the Popover is getting closed (not correct behavior), but
unsuccessfully, as its opener is rerendered.

Solution: The problem happens because of a new implementation in the
some controls, which sets "accessKey" property in the onBeforeRendering
logic, which invalidates the controls.
"accessKey" property should be added in the noInvalidationProps of
ToggleButtons' config, so the OverflowToolbar won't be
invalidated, due to its change, as it does not affect size.

BCP: 2280180459
SNOW: CS20240006967628
Change-Id: Id522a6790cabf6f6b55a5086f7e5a7b9c081a18c
CR-Id: 002075125900001002392024
(cherry picked from commit 8df676ffbb651a577aa07d79b03515ef749ae7ed)