[INTERNAL][FIX]sap.ui.rta: Invalidate cache on exit RTA without changes
It is now possible to exit RTA without saving changes; but not calling
the save method in this case prevented the FL cache invalidation.

Since choosing not to save the changes reverts them, with this fix the
save is always called to invalidate the cache. Also the reload check is
done before the save; because saving altered the check result.

Change-Id: Ib0eb5574773db8075dd8dba3d9eccba57d368392
BCP: 2280181842