[INTERNAL] sap.uxap.ObjectPage: Avoid double announcement of section titles
In addition to Ifbbf8f9d21d2fc2fb128a35b45c8a820c84640e6, titles of just promoted sections are also readout twice.
With this we make sure that title of promoted section is not being added to the label of the subsection as it is already added to the label of the parent section.
Better naming for the getter applied.

Change-Id: I63712c9ea82b73fb5540a0f438b5b6b7bf8dd349
BCP: 002075129500006802072022
BCP: 2270186107