[FIX] sap.MultiComboBox: Don't display clearIcon when there is no value
8e24738618d962d755af23defdae24fedd927bf5 exposed a bug in the MultiComboBox
that whenever the 'selectedKeys' property is set before the 'items' aggregation
the clearIcon is shown because the value is still not removed - because
the _synchronizeSelectedItemAndKey() method is still not executed.

Now we determine if we should display the clear icon on a later stage after
we sync the selected item and key and thus clear the initially set value.

BCP: 2270172883
Change-Id: Ice663bac1fb18cc1dac77ef556117f1b26bd1ffa