[FIX] sap.ui.unified.calendar.Month: fixed interval selection
* There was wrong date marking when there is 1 month displayed,
interval selection is enabled, the focused date is in currently
displayed month, and the start date is in the previous/next month.
In that case dates between start date and focused date are marked
initially as selected (interval between) even if there is no end date

* now this is fixed - when focused date is not in the same month as
start date, the dates between are not marked initially.

* test snippix: #778939

Change-Id: Icc057919dfb3acd9d1b2282aa0f7a45ce864a18d
BCP: 2270177246