[INTERNAL] ManagedObject: Move runAsOwner() from Component to ManagedObject
With #5511111 an asynchronous break out from the synchronous databinding lifecycle was introduced via the ListBase/GrowingEnablement.
ManagedObject(BindingSupport) was not ready for asynchronous owner-component propagation.
This change moves the runAsOwner() function to the ManagedObject.
Allowing to share code between Component and ManagedObjectBindingSupport and correctly scope the owner-component in the bindingInfo.factory().

CR-Id: 002075125800005415332022
BCP: 2280181714
Change-Id: I22bcfa211ba9ebe46e94a7914448de0b47709fb2