[INTERNAL] sap.m.ComboBoxBase: Open property deprecated
Due to a incorrect handling of the visibility of the 'open' property
of the ComboBoxBase the property was marked as public.
We need to deprecate the existing property in order to prevent
future misuse. A new 'open' private property
is introduced in the correct way in order to maintain the current behavior.

Change-Id: I531f567675d7093083ff02cfd5e23ff5d4755966
JIRA: BGSOFUIRILA-3491