[FIX] theming/Parameters: Fix parameter parsing
Prevent parsing of parameters based on a "dirty" DOM, where
an outdated link tag for a library CSS is still part of the
DOM and the CSS file for the current link tag of this library
already finished loading. This can possibly happen in case
Parameters.get is called after the browser finished loading
of CSS for the specific library and before the polling
mechanism of the themeChanged event is processed.

BCP: 002075129500009129202022
Change-Id: I02e152cf82ee0eadf9cf4faaca7be17ebe0d61b2
(cherry picked from commit I452008fae07e59ab9c3adf5fbc9f8bec6f158eef)
CR-Id: 002075125800005690192022