[FEATURE] sap.m.IllustratedMessage: New property introduced
New property called 'enableDefaultTitleAndDescription' is introduced.
It's purpose is to have the option to disable the default title and
description when we use the default illustration set and their values
are unset/empty.

It's default value is 'true'. This results in having the default texts.
We are preserving the previous default behavior for backward
compatibility.

If set to 'false' the default title or description won't be used even if
we use the default set and the values for title/description are empty.

Additionally we are now not rendering the title or description private
aggregations if their respective properties are empty. This means that
they won't occupy their designated space in the window rendering empty
fields.

JIRA: BGSOFUIPIRIN-5994
BCP: 2270181515
Change-Id: I0d79417ec61f6567e854c331b4c78549899fd126