[FEATURE] sap.m.ObjectStatus Add CSS class for in-word StatusText wrapping
A client cas requested that we add a CSS class which wuld allow the
text in ObjectStatus to wrap in-word, as opposed to the default
between-word wrapping.

Added a CSS rule that, when added to an ObjectStatus, allows the
text in its StatusText div to wrap anywhere when needed.

JIRA: BGSOFUIBALKAN-6443
Change-Id: I805b47874de4adb2f9ed3586c1d462f1ba5505e6