[FIX] sap.m.ObjectStatus Add CSS class for in-word StatusText wrapping
A client has requested that we add a CSS class which would allow the
text in ObjectStatus to wrap in-word, as opposed to the default
between-word wrapping.

Added a CSS rule that, when added to an ObjectStatus, allows the
text in its StatusText div to wrap anywhere when needed.

JIRA: BGSOFUIBALKAN-6443
BCP: N/A
CR-Id: 002075125800003675362023
Change-Id: I805b47874de4adb2f9ed3586c1d462f1ba5505e6
(cherry picked from commit 2314a65a9e02a6d350fef8b7da8b8fa9b259b43f)