[FIX] sap.ui.unified.calendar.Month: Fix week selection when Min & Max dates are set
Issue:
When min and/or max days are set, selecting a week that is partially out of
bounds will select all available dates insted of just the week.

Solution:
First/Last enabled date of week is no longer set to min/max day when it
is out of bounds.

BCP: 2280178657
Change-Id: Ife662ff4c02af353f2a67f86915566d0b816009b