[INTERNAL] Demo Kit: Code tag formatting fixed
Problem: The code tag in the subtitles was not rendered correctly.
It turned out that the code tag was being stripped out by the
XML2JSONUtils module.

Solution: Modified the removeHTMLTags functionality to allow
code tags based on configuration specified in the manifest.json and
changed the control in the view to use sap.m.FormattedText instead
of sap.m.Text to render the subtitle.

Jira: BGSOFUIPIRIN-5945
Change-Id: I149dacefb01f96d8dab2ce028112d0fa47a06cce