[FIX] sap.m.MessageView: List Item speech output improved
We used to set a metacharacter as a value of the info property
of a MessageListItem inside the MessageView, which resulted in a whitespace.
As the whitespace is a non-empty value of the info property,
the invisible text created for the aria-labelledby attribute of the StandardListItem was
including it, which resulted in an undesired speech output.
The info property used to be set as it was needed for applying an infoState to the
MessageListItem, which is not the case now, so this change removes it
in order to prevent the incorrect announcement of dots.


Change-Id: I593e6735e947e980f841eb27defd3aff5030d836
BCP: 2270182409