[INTERNAL][FIX] sap.ui.fl: Filter standard variants from cache key
The flex-related part of the chache key usually contains a list of all
active variant ids. However in the case of standard variants they might
not be known when the key is calculated.

If no changes are returned from the backend, the standard variant is
faked in the frontend after the VM control was initialized.
As a result there are two different scenarios:
1) App is loaded and the key is calculated before VM control is
initialized -> Key doesn't contain standard variant id
2) The cache key is created after a navigation and the VM control was
preloaded (e.g. by FE) -> Key exists
Both scenarios can occur within one session and thus create unnecessary
duplicate cache entries.
To fix this, standard variant ids are always filtered and not included
in the cache key.

Change-Id: I448e5f78c107a63f4412ffa35e9e80f2c78656f7
BCP: 2380001341