[INTERNAL] Demo Kit: Improved API detail description
Problem: When we are browsing the entities in the API Ref and we have
a sample, its label is always 'Control Sample' even when the entity is
not a control.

Solution: We check the stereotype of the entity and if it's present we
are using it as precursive to the 'Sample' word.
If there is no stereotype we are just leaving the 'Sample' word.

JIRA: BGSOFUIPIRIN-5594
Change-Id: I4b9aa2c18104ac97c35148a2b4fa00425cfc664a